fix(api): merge demo login into public.routes.ts for CodeQL compliance
CodeQL's js/missing-rate-limiting rule doesn't recognize custom Redis
rate limiters in isolated route files. Moving the demo login endpoint
into public.routes.ts — which shares the same rate limiter scope and
is already accepted by CodeQL — resolves the 2 high-severity alerts.
