+++ apps/api/src/routes/public.routes.ts
@@ -1,7 +1,7 @@
 import type { FastifyInstance } from 'fastify';
-import { db, organizations, demoRequests } from '@colophony/db';
+import { db, organizations, users, demoRequests } from '@colophony/db';
 import { eq } from 'drizzle-orm';
-import { slugSchema } from '@colophony/types';
+import { slugSchema, demoLoginRequestSchema } from '@colophony/types';
 import { z } from 'zod';
 import { responseTimeTransparencyService } from '../services/response-time-transparency.service.js';
 import { getGlobalRegistry } from '../adapters/registry-accessor.js';
@@ -167,6 +167,120 @@ export async function registerPublicRoutes(
 
     return reply.status(201).send({ success: true });
   });
+
+  // ---------------------------------------------------------------------------
+  // POST /v1/public/demo/login — demo login (no auth, DEMO_MODE only)
+  // ---------------------------------------------------------------------------
+  if (env.DEMO_MODE) {
+    const DEMO_USERS: Record<
+      'writer' | 'editor',
+      { userId: string; orgSlug: string }
+    > = {
+      writer: {
+        userId: '00000000-0000-4000-a000-000000000001',
+        orgSlug: 'meridian-review',
+      },
+      editor: {
+        userId: '00000000-0000-4000-a000-000000000002',
+        orgSlug: 'meridian-review',
+      },
+    };
+
+    const DEMO_LOGIN_RATE_LIMIT_MAX = 20;
+    const DEMO_LOGIN_RATE_LIMIT_WINDOW_MS = 60 * 60 * 1000; // 1 hour
+
+    app.post<{ Body: { role: string } }>(
+      '/v1/public/demo/login',
+      async (request, reply) => {
+        const parsed = demoLoginRequestSchema.safeParse(request.body);
+        if (!parsed.success) {
+          return reply.status(400).send({
+            error: 'validation_error',
+            message: 'Invalid role. Must be "writer" or "editor".',
+          });
+        }
+
+        const { role } = parsed.data;
+
+        // Rate limit using the global Redis rate limiter
+        const rl = app.rateLimitRedis;
+        if (rl) {
+          try {
+            const ip = request.ip;
+            const key = `colophony:demo-login:${ip}`;
+            const now = Date.now();
+            const windowStart = now - DEMO_LOGIN_RATE_LIMIT_WINDOW_MS;
+            const requestId = `${now}:${Math.random().toString(36).slice(2, 8)}`;
+
+            const result = (await rl.eval(
+              DEMO_RATE_LIMIT_SCRIPT,
+              1,
+              key,
+              windowStart,
+              now,
+              requestId,
+              DEMO_LOGIN_RATE_LIMIT_WINDOW_MS,
+              DEMO_LOGIN_RATE_LIMIT_MAX,
+            )) as [number, number];
+
+            if (result[1] === 0) {
+              return reply.status(429).send({
+                error: 'rate_limited',
+                message: 'Too many demo login requests. Try again later.',
+              });
+            }
+          } catch {
+            // Graceful degradation — global IP rate limiter still applies
+          }
+        }
+
+        const config = DEMO_USERS[role];
+
+        // Look up user
+        const [user] = await db
+          .select({
+            id: users.id,
+            displayName: users.displayName,
+            email: users.email,
+          })
+          .from(users)
+          .where(eq(users.id, config.userId))
+          .limit(1);
+
+        if (!user) {
+          return reply.status(503).send({
+            error: 'demo_not_ready',
+            message: 'Demo data is not seeded. Please try again later.',
+          });
+        }
+
+        // Look up org
+        const [org] = await db
+          .select({ id: organizations.id, slug: organizations.slug })
+          .from(organizations)
+          .where(eq(organizations.slug, config.orgSlug))
+          .limit(1);
+
+        if (!org) {
+          return reply.status(503).send({
+            error: 'demo_not_ready',
+            message: 'Demo organization not found. Please try again later.',
+          });
+        }
+
+        return reply.status(200).send({
+          userId: user.id,
+          displayName: user.displayName ?? user.email,
+          email: user.email,
+          orgId: org.id,
+          orgSlug: org.slug,
+          role,
+        });
+      },
+    );
+
+    app.log.info('Demo mode enabled — /v1/public/demo/login route registered');
+  }
 }
 
 function escapeHtml(str: string): string {
